Download Tiles of the Copernicus Digital Elevation Model.

REQUIREMENTS: This function requires aws cli installed in your system. To check that you have acces to the data you can run aws s3 ls --no-sign-request s3://copernicus-dem-90m/. Data sources:

Copernicus: https://spacedata.copernicus.eu/collections/copernicus-digital-elevation-model.

Open Data AWS: https://registry.opendata.aws/copernicus-dem/.

dem_download(
  polygon = NULL,
  resolution = 30,
  folder = "dem_tiles",
  output = "vrt",
  quiet = TRUE,
  sequential_download = FALSE
)

Arguments

polygon

(required, sf data frame) area to download, in CRS 4326. Default: NULL.

resolution

(required, integer) Either 30 or 90 meters resolution. Default: 30

folder

(optional, character string) Name of the folder where to download the tiles of the digital elevation model. Default: "dem_tiles".

output

(optional, output format) Either "vrt" for a file path to an virtual raster file with all downloaded tiles, or "raster" for a terra SpatRaster object. Default: "vrt"

quiet

(optional, logical) If TRUE all messages are suppressed. Set to FALSE for debugging. Default: TRUE

sequential_download

(optional, logical) If TRUE, tiles are downloaded sequentially. Set to TRUE for debugging. Default: FALSE

Examples

if(interactive()){

library(dem)
library(terra)

#subset one county
newton <- counties_sf[counties_sf$county == "Newton", ]

#download dem tiles for the county
dem <- dem_download(
 polygon = newton,
 resolution = 90,
 output = "raster"
)

#plot dem and polygons
plot(dem)
lines(terra::vect(newton))

}