Generate Topographic Indicators from a Digital Elevation Model

Generates the following topographic indicators:

topo_elevation

topo_slope: terrain slope in degrees

topo_northerness: level of terrain orientation to the north.

topo_easterness: level of terrain orientation to the east.

topo_tpi: difference between the value of a cell and the mean value of its 8 surrounding cells.

topo_tri: Terrain Ruggedness Index is the mean of the absolute differences between the value of a cell and its 8 surrounding cells.

topo_roughness: difference between the maximum and the minimum value of a cell and its 8 surrounding cells.

topo_flowdir: direction of the greatest drop in elevation, or the smallest rise if all neighbors are higher.

dem_indicators(
  raster = NULL,
  dem = NULL,
  indicators = c("slope", "aspect", "TPI", "TRI", "roughness", "flowdir")
)

Arguments

raster

(optional, SpatRaster) A "terra" SpatRaster object to be used as a template to crop and resample the Digital Elevation Model. If not provided, topographic indicators are computed for the complete set of tiles available in the dem argument. Default: NULL

dem

(required, SpatRaster or path to virtual raster file) Digital Elevation Model generated by dem_download() provided as a SpatRaster or as a path to a vrt file. Default: NULL

indicators

(optional, character vector) Names of the topographic indicators to generate. Argument v of terra::terrain. These names are converted to lowercase and attached the prefix "topo_" in the output raster file. Default: c("slope", "aspect", "TPI", "TRI", "roughness", "flowdir").

Value

Multilayer SpatRaster object with the selected indicators. If raster is provided, then the output object contains all raster layers in raster as well.

Examples

if(interactive()){

library(dem)
library(terra)

#subset one county
newton <- counties_sf[counties_sf$county == "Newton", ]

#download dem tiles for the county
dem <- dem_download(
 polygon = newton,
 resolution = 90,
 output = "raster"
)

#plot dem and polygons
plot(dem)
lines(terra::vect(newton))

#compute topographic indicators
topography <- dem_indicators(
 dem = dem
)

#plot topography
plot(topography)

}